Стилі для dropdown виглядають вже сносно. Розділив серверні та клієнтскі компоненти з Profile, по двом файлам, які відповідно містять серверні і клієнтьскі компоненти
