+++ app/components/Header/index.tsx
@@ -1,14 +1,15 @@
+import {
+	NavigationMenu,
+	NavigationMenuList
+} from '@/components/ui/navigation-menu';
 import Logo from './Logo';
 import Navigation from './Navigation';
-import ProfileSection from './Profile';
-import { NavigationMenu, NavigationMenuItem, NavigationMenuList } from '@radix-ui/react-navigation-menu';
-
-
+import ProfileSection from './Profile-server';
 
 export default function Header() {
 	return (
-		<NavigationMenu className='header-cont'>
-			<NavigationMenuList className='header'>
+		<NavigationMenu className='header-cont h-header bg-navLayout'>
+			<NavigationMenuList className='flex justify-between w-screen py-4 px-8'>
 				<Logo />
 				<Navigation />
 				<ProfileSection />
